Change DLFUCache_perf.py walk() to take variance, increase size to 1000.
The stochastic walk() standard deviation of "distance walked in N
iterations" is sigma*sqrt(N), so the impact on cache hitrates as you
increase cache size is more constant if you specfy the walk variance
as a multiple of the cache size.

Increasing the cache size to 1000 should give less noisy results.