Add ARCCache to DLFUCache_perf.py and tune synthetic loads.
Make walk() wrap instead of reflect at minv and maxv boundaries. Also
make it start by default at MAXK/2 in the middle of the range
boundaries. This makes it not "sticky" at the boundaries and ensures
it starts offset from other loads in mixed().

Make scan() wrap past maxv instead of reset to minv, and make it
always return minv <= v < maxv. This has no practical effect but is a
bit nicer.

Make jump() step default to 4.0 instead of 2.0 so that there is less
overlap between steps.

Make mixed() start the jump() with an offset of 4*size so that it
doesn't start overlapped with the expo() load pattern.

Make alltests() do mixed(N/4) instead of mixed(N/2) to compensate for
all the expo()/jump()/walk() patterns not starting overlapped any more.

Add an alltests() run for ARCCache for comparison.

Update the output data file.