Add binheap PQueue implementation.
Simplify __init__() to init data from the queue.

Rename PQueueHeapq to PQueueHeapqD and update docstring to mention
that it uses Python's heapq with deletion markers.

Add new PQueueHeapq using a binheap with entries containing their index.

Update PQueue_perf.py so mhit() and miss() better reflect actual
DLFUCache queue operations and always ensure the keys in the queue are
in the range [0:N]. Change the tests to take advantage of this and
work when the number of iterations is larger than the queue size.

Change the min size tested from 4K to 1K to better reflect sizes used
by DLFUCache_perf.py, and increase the number of iterations from 1000
to 2^14 to better show the cost of using deletion markers.

Add HeapqD to the list of queues tested.

Update PQueue_perf.out with the new results.