Adjust wave() sequence generator for better behaviour.
This changes the expo median default from size/3 to size/2 and reduces the
sweep step from 0.5 to 0.25. This gives a theoretical peak hit rate for larger
caches of 75%, because 4*size fetches after an entry is first fetched at the
leading edge of the wave, the wave will have swept past, so the entry is now
on the wave's tail and not worth caching. This means each entry gets avg 4
fetches, and the first is always a miss. A median of size/2 gives our target
50% hit rate.