Add a version of DLFU that uses a PIDController for expiry.
This removes the priority queues, which are O(ln N) for cache misses, and
replaces them with trimming the cache back to 95% full every time it gets
full. The trimming is done with a simple linear scan through all the entries
and evicting them from the cache or metadata if their count is less than a
threshold, where the theshold is driven by a PIDController targeting 95% full.

This makes all cache operations amortized O(1), and might be cheaper than
maintaining priority queues, and should perform well provided the
PIDController can be tuned to work well.